requirements.txt
numpy>=1.21.0
sounddevice>=0.4.0
openai-whisper>=20231117
scipy>=1.7.0
pyttsx3>=2.90
google-generativeai>=0.3.0
python-dotenv>=0.19.0
langgraph>=0.0.20